pyproject.toml
[project]
name = "config-sentinel"
version = "0.1.0"
description = "A Python library for seamless configuration management, supporting JSON, YAML, and TOML formats with real-time file watching, robust validation, and intuitive dot-notation access."
readme = "README.md"
authors = [
    { name = "Jai Brown", email = "jaintp.dev@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10.5",
    "pyyaml>=6.0.2",
    "tomli>=2.2.1",
    "tomli-w>=1.2.0",
    "watchdog>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "coverage-badge>=1.1.2",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
]

[tool.coverage.run]
omit = [
    'tests/*'
]

[tool.pytest.ini_options]
addopts = "--cov=eventbus --cov-report term-missing"
